Image_watermark_Adder/requirements.txt
streamlit
pillow
